Ensure the mounted config directory contains /etc/elasticsearch
The files in /etc/elasticsearch are copied into the mounted
configuration directory and then the directory is symlinked.
This also allows users to pre-seed configuration files with
their own changes.

We also need to retry the commands as there are times where
`docker exec` returns a non-zero value for a command that
should run successfully. This is probably some weird docker
race condition, so retrying seems appropriate.

Closes #16