fix: use a temporary container to force set permissions on elasticsearch config and data
Previously, we would retry setting the permissions on the running container, resulting in ES sometimes coming up in a state where it wasn't ready, but always resulting in longer boot times.

We now use an ephemeral container to set permissions for the config and data directories.

This should also increase compatibility with newer versions of elasticsearch as we won't be impacted by ES immediately crashing the container on permission error.